Update OAuth clients
Add THINC-IT as an OAuth client. Update rest-api to only have client-credentials grant. Update expiry times of login tokens (ManagementPortal) to 30 minutes for access, 60 minutes for refresh. Update expiry times for app tokens to 12 hours for access token (to limit amount of requests for tokens), and 60 days for refresh token. This assumes a phone is offline for no more than 60 days. After this a patient would have to acquire a new qr code to login again.