Support for clients authorized by client_credentials
Instead of relying on getting the current user and then getting their authorities, rely on more general Spring framework SecurityContext and Authorization to check authorities in SourceService, SubjectService and ProjectService. Roles do not make sense in the case of a client_credentials, so there we just return an empty list. In AccountResource, return NOT_FOUND if we could not find the user (like would be the case with client_credentials). INTERNAL_SERVER_ERROR would indicate something being wrong. Nothing is wrong, there just is no user, so return NOT_FOUND.