Fix for hashCode in SensorData objects
If hashCode is only based on the id field, SensorData's without an id will have the same hash. This is a problem when POSTing a set of new SensorData's, as they have a null id field and as a result will all look the same to a HashSet. So only one SensorData can be in the HashSet in that case.