Change updateParticipantRoles
Change in updateParticipantRoles, with the previous version the INACTIVE_PARTICIPANT role was always added, even if it was not necessary. While this is not a big deal, it 'pollutes' the revision log with a change in the project when a change in a subject triggered this role to be added.